+++ cats/cats-dynomite/src/main/java/com/netflix/spinnaker/cats/dynomite/cache/DynomiteCache.java
@@ -15,10 +15,13 @@
  */
 package com.netflix.spinnaker.cats.dynomite.cache;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.common.collect.Lists;
+import com.google.common.hash.Hashing;
 import com.netflix.dyno.connectionpool.exception.DynoException;
 import com.netflix.spinnaker.cats.cache.CacheData;
+import com.netflix.spinnaker.cats.cache.DefaultCacheData;
 import com.netflix.spinnaker.cats.dynomite.DynomiteClientDelegate;
 import com.netflix.spinnaker.cats.redis.cache.AbstractRedisCache;
 import com.netflix.spinnaker.cats.redis.cache.RedisCacheOptions;
@@ -28,18 +31,58 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import redis.clients.jedis.exceptions.JedisException;
 
+import java.io.IOException;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
+import java.util.Set;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
+
 public class DynomiteCache extends AbstractRedisCache {
+
+  public interface CacheMetrics {
+    default void merge(String prefix,
+                       String type,
+                       int itemCount,
+                       int relationshipCount,
+                       int hashMatches,
+                       int hashUpdates,
+                       int saddOperations,
+                       int hmsetOperations,
+                       int expireOperations,
+                       int delOperations) {
+      // noop
+    }
+
+    default void evict(String prefix,
+                       String type,
+                       int itemCount,
+                       int delOperations,
+                       int sremOperations) {
+      // noop
+    }
+
+    default void get(String prefix,
+                     String type,
+                     int itemCount,
+                     int requestedSize,
+                     int relationshipsRequested,
+                     int hmgetAllOperations) {
+      // noop
+    }
+
+    class NOOP implements CacheMetrics {}
+  }
 
   private final Logger log = LoggerFactory.getLogger(getClass());
 
@@ -48,8 +91,11 @@ public class DynomiteCache extends AbstractRedisCache {
     .withDelay(500, TimeUnit.MILLISECONDS)
     .withMaxRetries(3);
 
+  private final CacheMetrics cacheMetrics;
+
   public DynomiteCache(String prefix, DynomiteClientDelegate dynomiteClientDelegate, ObjectMapper objectMapper, RedisCacheOptions options, CacheMetrics cacheMetrics) {
-    super(prefix, dynomiteClientDelegate, objectMapper, options, cacheMetrics);
+    super(prefix, dynomiteClientDelegate, objectMapper, options);
+    this.cacheMetrics = cacheMetrics == null ? new CacheMetrics.NOOP() : cacheMetrics;
   }
 
   @Override
@@ -58,22 +104,20 @@ public class DynomiteCache extends AbstractRedisCache {
       return;
     }
 
-    AtomicInteger keysWritten = new AtomicInteger();
     AtomicInteger relationships = new AtomicInteger();
+    AtomicInteger hmsetOperations = new AtomicInteger();
     AtomicInteger saddOperations = new AtomicInteger();
-    AtomicInteger setOperations = new AtomicInteger();
     AtomicInteger expireOperations = new AtomicInteger();
     AtomicInteger delOperations = new AtomicInteger();
     AtomicInteger skippedWrites = new AtomicInteger();
     AtomicInteger hashesUpdated = new AtomicInteger();
 
-    final Map<String, String> hashes = getHashes(type, items);
     redisClientDelegate.withCommandsClient(client -> {
       for (CacheData item : items) {
-        MergeOp op = buildMergeOp(type, item, hashes);
+        MergeOp op = buildHashedMergeOp(type, item);
         skippedWrites.addAndGet(op.skippedWrites);
 
-        if (op.keysToSet.isEmpty()) {
+        if (op.valuesToSet.isEmpty()) {
           continue;
         }
 
@@ -81,19 +125,13 @@ public class DynomiteCache extends AbstractRedisCache {
           .with(redisRetryPolicy)
           .onFailure(failure -> {
             log.error("Encountered repeated failures while setting " + type + " cache data for " + item.getId(), failure);
-            op.hashesToSet.keySet().forEach(it -> {
-              String hashKey = hashKey(hashesId(type), it);
-              client.del(hashKey);
-              delOperations.incrementAndGet();
-            });
+            client.del(itemHashesId(type, item.getId()));
+            delOperations.incrementAndGet();
             throw new RuntimeException("Failed running caching agent", failure);
           })
           .run(() -> {
-            for (int i = 0; i < op.keysToSet.size(); i = i + 2) {
-              client.set(op.keysToSet.get(i), op.keysToSet.get(i + 1));
-              setOperations.incrementAndGet();
-              keysWritten.incrementAndGet();
-            }
+            client.hmset(itemId(type, item.getId()), op.valuesToSet);
+            hmsetOperations.incrementAndGet();
 
             if (!op.relNames.isEmpty()) {
               client.sadd(allRelationshipsId(type), op.relNames.toArray(new String[op.relNames.size()]));
@@ -102,10 +140,8 @@ public class DynomiteCache extends AbstractRedisCache {
             }
 
             if (item.getTtlSeconds() > 0) {
-              for (String key : op.keysToSet) {
-                client.expire(key, item.getTtlSeconds());
-                expireOperations.incrementAndGet();
-              }
+              client.expire(itemId(type, item.getId()), item.getTtlSeconds());
+              expireOperations.incrementAndGet();
             }
 
             client.sadd(allOfTypeReindex(type), item.getId());
@@ -114,11 +150,11 @@ public class DynomiteCache extends AbstractRedisCache {
             saddOperations.incrementAndGet();
 
             if (!op.hashesToSet.isEmpty()) {
-              for (Entry<String, String> hashEntry : op.hashesToSet.entrySet()) {
-                client.setex(hashKey(hashesId(type), hashEntry.getKey()), getHashExpiry(), hashEntry.getValue());
-                setOperations.incrementAndGet();
-                hashesUpdated.incrementAndGet();
-              }
+              client.hmset(itemHashesId(type, item.getId()), op.hashesToSet);
+              hmsetOperations.incrementAndGet();
+              client.expire(itemHashesId(type, item.getId()), getHashExpiry());
+              expireOperations.incrementAndGet();
+              hashesUpdated.addAndGet(op.hashesToSet.size());
             }
           });
       }
@@ -128,15 +164,11 @@ public class DynomiteCache extends AbstractRedisCache {
       prefix,
       type,
       items.size(),
-      keysWritten.get(),
       relationships.get(),
       skippedWrites.get(),
       hashesUpdated.get(),
       saddOperations.get(),
-      setOperations.get(),
-      0,
-      0,
-      0,
+      hmsetOperations.get(),
       expireOperations.get(),
       delOperations.get()
     );
@@ -144,14 +176,6 @@ public class DynomiteCache extends AbstractRedisCache {
 
   @Override
   protected void evictItems(String type, List<String> identifiers, Collection<String> allRelationships) {
-    List<String> delKeys = new ArrayList<>((allRelationships.size() + 1) * identifiers.size());
-    for (String id : identifiers) {
-      for (String rel : allRelationships) {
-        delKeys.add(relationshipId(type, id, rel));
-      }
-      delKeys.add(attributesId(type, id));
-    }
-
     AtomicInteger delOperations = new AtomicInteger();
     AtomicInteger sremOperations = new AtomicInteger();
 
@@ -166,10 +190,10 @@ public class DynomiteCache extends AbstractRedisCache {
           sremOperations.incrementAndGet();
         }
 
-        for (String key : delKeys) {
-          client.del(key);
+        for (String id : identifiers) {
+          client.del(itemId(type, id));
           delOperations.incrementAndGet();
-          client.del(hashKey(hashesId(type), key));
+          client.del(itemHashesId(type, id));
           delOperations.incrementAndGet();
         }
       }));
@@ -178,32 +202,153 @@ public class DynomiteCache extends AbstractRedisCache {
       prefix,
       type,
       identifiers.size(),
-      delKeys.size(),
-      delKeys.size(),
       delOperations.get(),
-      0,
       sremOperations.get()
     );
   }
 
   @Override
-  protected List<String> getHashValues(List<String> hashKeys, String hashesId) {
-    final List<String> hashValues = new ArrayList<>(hashKeys.size());
-    redisClientDelegate.withCommandsClient(c -> {
-      // TODO rz - Dynomite mget perf is O(n^2), once fixed we can go to mget.
-      for (String hashKey : hashKeys) {
-        hashValues.add(c.get(hashKey(hashesId, hashKey)));
+  protected Collection<CacheData> getItems(String type, List<String> ids, List<String> knownRels) {
+    Map<String, Map<String, String>> rawItems = new HashMap<>();
+    int hmgetAllOperations = redisClientDelegate.withCommandsClient(c -> {
+      int ops = 0;
+      for (String id : ids) {
+        Map<String, String> item = c.hgetAll(itemId(type, id));
+        if (item != null && !item.isEmpty()) {
+          rawItems.put(id, item);
+          ops++;
+        }
       }
+      return ops;
     });
-    return hashValues;
+
+    Collection<CacheData> results = new ArrayList<>(ids.size());
+    for (Map.Entry<String, Map<String, String>> rawItem : rawItems.entrySet()) {
+      CacheData item = extractHashedItem(rawItem.getKey(), rawItem.getValue(), knownRels);
+      if (item != null) {
+        results.add(item);
+      }
+    }
+
+    cacheMetrics.get(prefix, type, results.size(), ids.size(), knownRels.size(), hmgetAllOperations);
+    return results;
   }
 
-  private String hashKey(String hashesId, String key) {
-    return hashesId + ":" + key;
+  private CacheData extractHashedItem(String id, Map<String, String> values, List<String> knownRels) {
+    if (values == null) {
+      return null;
+    }
+
+    try {
+      final Map<String, Object> attributes;
+      if (values.get("attributes") != null) {
+        attributes = objectMapper.readValue(values.get("attributes"), ATTRIBUTES);
+      } else {
+        attributes = null;
+      }
+      final Map<String, Collection<String>> relationships = new HashMap<>();
+      for (Map.Entry<String, String> value : values.entrySet()) {
+        // TODO rz - Get relationships individually? Is this actually an optimization?
+        if (value.getKey().equals("attributes") || value.getKey().equals("id") || !knownRels.contains(value.getKey())) {
+          continue;
+        }
+        Collection<String> deserializedRel = objectMapper.readValue(value.getValue(), RELATIONSHIPS);
+        relationships.put(value.getKey(), deserializedRel);
+      }
+      return new DefaultCacheData(id, attributes, relationships);
+    } catch (IOException deserializationException) {
+      throw new RuntimeException("Deserialization failed", deserializationException);
+    }
+  }
+
+  private static class MergeOp {
+    final Set<String> relNames;
+    final Map<String, String> valuesToSet;
+    final Map<String, String> hashesToSet;
+    final int skippedWrites;
+
+    public MergeOp(Set<String> relNames, Map<String, String> valuesToSet, Map<String, String> hashesToSet, int skippedWrites) {
+      this.relNames = relNames;
+      this.valuesToSet = valuesToSet;
+      this.hashesToSet = hashesToSet;
+      this.skippedWrites = skippedWrites;
+    }
+  }
+
+  private boolean hashCheck(Map<String, String> hashes, String id, String serializedValue, Map<String, String> updatedHashes) {
+    if (options.isHashingEnabled()) {
+      final String hash = Hashing.sha1().newHasher().putString(serializedValue, UTF_8).hash().toString();
+      final String existingHash = hashes.get(id);
+      if (hash.equals(existingHash)) {
+        return true;
+      }
+      updatedHashes.put(id, hash);
+    }
+    return false;
+  }
+
+  private MergeOp buildHashedMergeOp(String type, CacheData cacheData) {
+    int skippedWrites = 0;
+    final String serializedAttributes;
+    try {
+      if (cacheData.getAttributes().isEmpty()) {
+        serializedAttributes = null;
+      } else {
+        serializedAttributes = objectMapper.writeValueAsString(cacheData.getAttributes());
+      }
+    } catch (JsonProcessingException serializationException) {
+      throw new RuntimeException("Attribute serialization failed", serializationException);
+    }
+
+    final Map<String, String> hashes = getHashes(type, cacheData);
+
+    final Map<String, String> hashesToSet = new HashMap<>();
+    final Map<String, String> valuesToSet = new HashMap<>();
+    if (serializedAttributes != null && hashCheck(hashes, attributesId(type, cacheData.getId()), serializedAttributes, hashesToSet)) {
+      skippedWrites++;
+    } else if (serializedAttributes != null) {
+      valuesToSet.put("attributes", serializedAttributes);
+    }
+
+    if (!cacheData.getRelationships().isEmpty()) {
+      for (Map.Entry<String, Collection<String>> relationship : cacheData.getRelationships().entrySet()) {
+        final String relationshipValue;
+        try {
+          relationshipValue = objectMapper.writeValueAsString(new LinkedHashSet<>(relationship.getValue()));
+        } catch (JsonProcessingException serializationException) {
+          throw new RuntimeException("Relationship serialization failed", serializationException);
+        }
+        if (hashCheck(hashes, relationshipId(type, cacheData.getId(), relationship.getKey()), relationshipValue, hashesToSet)) {
+          skippedWrites++;
+        } else {
+          valuesToSet.put(relationship.getKey(), relationshipValue);
+        }
+      }
+    }
+
+    return new MergeOp(cacheData.getRelationships().keySet(), valuesToSet, hashesToSet, skippedWrites);
+  }
+
+  private Map<String, String> getHashes(String type, CacheData item) {
+    if (isHashingDisabled(type)) {
+      return Collections.emptyMap();
+    }
+
+    return redisClientDelegate.withCommandsClient(c -> {
+      return c.hgetAll(itemHashesId(type, item.getId()));
+    });
   }
 
   private int getHashExpiry() {
     // between 1 and 3 hours; boundary is exclusive
     return (int) Duration.ofMinutes(ThreadLocalRandom.current().nextInt(60, 4 * 60)).getSeconds();
+  }
+
+  private String itemId(String type, String id) {
+    return String.format("%s:%s:item:%s", prefix, type, id);
+  }
+
+  private String itemHashesId(String type, String id) {
+    return String.format("%s:%s:hashes:%s", prefix, type, id);
   }
 }
